Add kubejobs controller and actuator, and includes redis metric source
The KubeJobs plugin get progress metrics from redis and provide pods
autoscaling over a k8s cluster that is running a batch of tasks (k8s-job).

The controller is similar to progress error using monasca-openstack. It
check the error (difference between proportional elapsed time and  pro-
portional application progress, based on how many tasks are done until
that time comparing with the total of tasks) and decides if the resour-
ces must be {in/de}creased considering max and min limits and variation
tolerance.

The actuator basicaly change the number of replicas deployed to this job
in k8s, where replicas are containers that can process any enqueued task.
To authenticate with k8s cluster and take actions like that, it is neces-
sary to indicate the k8s_manifest (yaml) in the controller.cfg.

The Redis metric_source is a plugin that make possible to consume metrics
from a Redis queue.